Extract components and add storybook
